♻️ refactor(formatter): rename and replace formater with formatter
